third_party/blink/web_tests/http/tests/xmlhttprequest/sync-xhr-revalidate-after-async-xhr-expected.txt

CONSOLE WARNING: Synchronous XMLHttpRequest on the main thread is deprecated because of its detrimental effects to the end user's experience. For more help, check https://xhr.spec.whatwg.org/.
Sync XHR should not turn async due to revalidation.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS xhr2SendIsReturned is false
PASS xhr1.responseText is not xhr2.responseText
PASS successfullyParsed is true

TEST COMPLETE
